moved components directly into pages that only had one container component
