+++ client/src/App.jsx
@@ -9,14 +9,17 @@ import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons';
 library.add(fas, faGithub, faLinkedin);
 
 import '../build/styles/index.css';
+import RootPage from './pages/RootPage/RootPage.jsx';
 
 const App = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route exact path='/' element={<LandingPage />} />
-        <Route path='/login' element={<LoginPage />} />
-        <Route path='/dashboard' element={<DashboardPage />} />
+        <Route path='/' element={<RootPage />}>
+          <Route exact path='/' element={<LandingPage />} />
+          <Route path='login' element={<LoginPage />} />
+        </Route>
+        <Route path='dashboard' element={<DashboardPage />} />
       </Routes>
     </BrowserRouter>
   );
