+++ client/src/pages/RootPage/components/Footer.jsx
