+++ client/src/pages/dashboardPage/DashboardPage.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import DashNav from './containers/DashNav';
 import BrokersContainer from './containers/BrokersContainer';
 import { v4 as uuidv4 } from 'uuid';
-import Footer from '../landingPage/containers/Footer';
+import Footer from '../RootPage/components/Footer';
 const DashboardPage = () => {
   const [username, setUsername] = useState('Demo User');
   const [connectionString, setConnectionString] = useState('prometheus:9090');
