+++ client/src/pages/loginPage/LoginPage.jsx
@@ -1,8 +1,7 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import Navbar from './containers/NavBar';
-import LoginBox from './containers/LoginBox';
-import Footer from '../landingPage/containers/Footer';
+import TextField from '../RootPage/components/TextField';
+import Button from '../RootPage/components/Button';
 
 const LoginPage = () => {
   const [username, setUsername] = useState(null);
@@ -28,13 +27,32 @@ const LoginPage = () => {
   };
   return (
     <div id='login-page' className='pages'>
-      <Navbar />
-      <LoginBox
-        setUsername={setUsername}
-        setPassword={setPassword}
-        handleSend={handleSend}
-      />
-      <Footer />
+      <div id='login-container'>
+        <TextField
+          id='username'
+          label='Username: '
+          onChange={setUsername}
+          isRequired
+        />
+        <TextField
+          id='password'
+          label='Password: '
+          onChange={setPassword}
+          isRequired
+        />
+        <div id='account-buttons'>
+          <Button id='login' onPress={() => handleSend('login')}>
+            Login
+          </Button>
+          <Button id='signup' onPress={() => handleSend('signup')}>
+            Sign up
+          </Button>
+        </div>
+        <hr />
+        <Button id='demo-button' onClick={() => navigate('/dashboard')}>
+          See a Demo!
+        </Button>
+      </div>
     </div>
   );
 };
